43_Nemo/requirement.txt
matplotlib.pyplot
numpy
scikit-learn
ipython
